changing ts unit formatting..

+++ lib/tsim/include/TSim/Simulation/Testbench.h
@@ -12,7 +12,7 @@
     simtb->GetUIntParam (Testbench::PARAMETER, #x);
 
 /* general */
-#define TO_SPEC_TIMEUNIT(simtime) (simtime / 1000)
+#define TO_SPEC_TIMEUNIT(simtime) ((double)simtime / 1000)
 #define TO_SIM_TIMEUNIT(spectime) (spectime * 1000)
 
 #include <TSim/Base/Metadata.h>
